moneymong-171 feat: ledgerdetail package refactor
