+++ core/network/src/main/java/com/moneymong/moneymong/network/request/ledgerdetail/LedgerTransactionDetailRequest.kt
@@ -1,4 +1,4 @@
-package com.moneymong.moneymong.network.request.ledger
+package com.moneymong.moneymong.network.request.ledgerdetail
 
 data class LedgerTransactionDetailRequest(
     val storeInfo: String,
