+++ core/network/src/main/java/com/moneymong/moneymong/network/response/ledgerdetail/LedgerTransactionDetailResponse.kt
@@ -1,4 +1,7 @@
-package com.moneymong.moneymong.network.response.ledger
+package com.moneymong.moneymong.network.response.ledgerdetail
+
+import com.moneymong.moneymong.network.response.ledger.DocumentImageURL
+import com.moneymong.moneymong.network.response.ledger.ReceiptImageURL
 
 data class LedgerTransactionDetailResponse(
     val id: Int,
