+++ data/src/main/java/com/moneymong/moneymong/data/di/RepositoryModule.kt
@@ -1,8 +1,10 @@
 package com.moneymong.moneymong.data.di
 
 import com.moneymong.moneymong.data.repository.ledger.LedgerRepositoryImpl
+import com.moneymong.moneymong.data.repository.ledgerdetail.LedgerDetailRepositoryImpl
 import com.moneymong.moneymong.data.repository.ocr.OCRRepositoryImpl
 import com.moneymong.moneymong.domain.repository.ledger.LedgerRepository
+import com.moneymong.moneymong.domain.repository.ledgerdetail.LedgerDetailRepository
 import com.moneymong.moneymong.domain.repository.ocr.OCRRepository
 import dagger.Binds
 import dagger.Module
@@ -22,4 +24,9 @@ interface RepositoryModule {
     fun provideLedgerRepository(
         ledgerRepositoryImpl: LedgerRepositoryImpl
     ): LedgerRepository
+
+    @Binds
+    fun provideLedgerDetailRepository(
+        ledgerDetailRepositoryImpl: LedgerDetailRepositoryImpl
+    ): LedgerDetailRepository
 }
