+++ data/src/main/java/com/moneymong/moneymong/data/mapper/ledgerdetail/LedgerDetailMapper.kt
@@ -3,9 +3,9 @@ package com.moneymong.moneymong.data.mapper.ledgerdetail
 import com.moneymong.moneymong.domain.param.ledgerdetail.LedgerDocumentParam
 import com.moneymong.moneymong.domain.param.ledgerdetail.LedgerReceiptParam
 import com.moneymong.moneymong.domain.param.ledgerdetail.LedgerTransactionDetailParam
-import com.moneymong.moneymong.network.request.ledger.LedgerDocumentRequest
-import com.moneymong.moneymong.network.request.ledger.LedgerReceiptRequest
-import com.moneymong.moneymong.network.request.ledger.LedgerTransactionDetailRequest
+import com.moneymong.moneymong.network.request.ledgerdetail.LedgerDocumentRequest
+import com.moneymong.moneymong.network.request.ledgerdetail.LedgerReceiptRequest
+import com.moneymong.moneymong.network.request.ledgerdetail.LedgerTransactionDetailRequest
 
 fun LedgerReceiptParam.toRequest() =
     LedgerReceiptRequest(
