+++ domain/src/main/java/com/moneymong/moneymong/domain/usecase/ledgerdetail/UpdateLedgerTransactionDetailUseCase.kt
@@ -4,11 +4,12 @@ import com.moneymong.moneymong.domain.base.BaseUseCase
 import com.moneymong.moneymong.domain.entity.ledger.LedgerTransactionDetailEntity
 import com.moneymong.moneymong.domain.param.ledgerdetail.LedgerTransactionDetailParam
 import com.moneymong.moneymong.domain.repository.ledger.LedgerRepository
+import com.moneymong.moneymong.domain.repository.ledgerdetail.LedgerDetailRepository
 import javax.inject.Inject
 
 class UpdateLedgerTransactionDetailUseCase @Inject constructor(
-    private val ledgerRepository: LedgerRepository
+    private val ledgerDetailRepository: LedgerDetailRepository
 ): BaseUseCase<LedgerTransactionDetailParam, Result<LedgerTransactionDetailEntity>>() {
     override suspend fun invoke(data: LedgerTransactionDetailParam): Result<LedgerTransactionDetailEntity> =
-        ledgerRepository.updateLedgerTransactionDetail(data)
+        ledgerDetailRepository.updateLedgerTransactionDetail(data)
 }
